+++ packages/api/src/index.ts
@@ -1,7 +1,7 @@
 export * from './constituents';
 export * from './datasets';
 export * from './definitions';
+export * from './enrichments';
 export * from './objects';
 export * from './organizations';
 export * from './provenance-events';
-export * from './provenance-events-searcher';
